+++ src/main/kotlin/game/GameStateSender.kt
@@ -3,6 +3,6 @@ package game
 interface GameStateSender {
     fun sendScore(playerNameToScore: Map<String, Int>)
     fun sendQuestion(question: Triple<String, List<String>, Int>)
-    fun sendRemainingTime(remainingTime: Double)
+    fun sendRemainingTime(remainingTime: Int)
     fun sendFinish()
 }
